+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+- [#604](https://github.com/persistenceOne/pstake-native/pull/604) Create an empty deposit when a host chain is registered.
+
 ## [v2.2.1] - 2023-07-27
 
 ### Improvements